APOLLO-3946 Renamed to singular namespace
